+++ spec/features/users_can_view_a_table_of_project_openers_spec.rb
@@ -13,22 +13,22 @@ RSpec.feature "Users can view project openers in table form" do
     end
 
     context "when a project has all conditions met" do
-      it "shows the correct tag" do
+      it "shows the correct value" do
         tasks_data = create(:conversion_tasks_data, conditions_met_confirm_all_conditions_met: true)
         _project = create(:conversion_project, tasks_data: tasks_data, conversion_date: Date.new(2023, 1, 1), conversion_date_provisional: false)
 
         visit confirmed_all_opening_projects_path(1, 2023)
-        expect(page).to have_css("strong.govuk-tag--turquoise", text: "confirmed")
+        expect(page).to have_content("Yes")
       end
     end
 
     context "when a project does not have all conditions met" do
-      it "shows the correct tag" do
+      it "shows the correct value" do
         tasks_data = create(:conversion_tasks_data, conditions_met_confirm_all_conditions_met: nil)
         _project = create(:conversion_project, tasks_data: tasks_data, conversion_date: Date.new(2023, 1, 1), conversion_date_provisional: false)
 
         visit confirmed_all_opening_projects_path(1, 2023)
-        expect(page).to have_css("strong.govuk-tag--blue", text: "unconfirmed")
+        expect(page).to have_content("Not yet")
       end
     end
   end