+++ CHANGELOG.md
@@ -15,6 +15,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- The opening view no longer shows the status of the direction to transfer or
+  trust modification order tasks.
+- The all conditions met column on the opening view is no longer shown as a
+  coloured tag and is one of Yes or Not yet.
+
 ### Fixed
 
 ## [Release 30][release-30]